+++ domain/src/main/java/com/findapple/findapple/domain/features/post/entity/Post.kt
@@ -2,10 +2,9 @@ package com.findapple.findapple.domain.features.post.entity
 
 import com.findapple.findapple.domain.entity.Location
 import com.findapple.findapple.domain.entity.User
-import java.util.*
 
 data class Post(
-    val id: Int,
+    val id: Long,
     val title: String,
     val user: User,
     val detailInfo: String,
